ADD Hydration
When returning data, we often want to include more data bundled with
it. For that we use hydration.

Example:
When someone is requesting a list of projects, we want to send owner
for each project with the request, to save the client another
requests (downloading the owners one by one, or bundled).

Hydration works in 2 phases:
 - first it will gather all keys which should be loaded from datastore
 - then it will load the keys and hydrate results with loaded entities