FIX AuthorizationInterceptor should acces the principal lazily
Since principal is supposed to be @RequestScoped, we cannot always
access it eagerly. The Authorization interceptor can be invoked in
other stages (like when Guice is injecting members using setters) and
thus cannot be expected to get current principal.

All methods annotated with Guarding annotations should be called only
when the Principal is available (ie. during some request) and from now
on the principal will be required only when a method is annotated and
authorization interceptor will run some guard on it.